(TRX): contract address modal don't show if no url available
